[ADD/#54] 친구추가 리스트 구현

+++ data/src/main/java/com/example/data/repository/OnboardingRepositoryImpl.kt
@@ -57,7 +57,7 @@ class OnboardingRepositoryImpl @Inject constructor(
 
     override suspend fun postToGetFriendList(
         request: RequestOnboardingListModel,
-        page: Long
+        page: Int
     ): FriendList? {
         return onboardingDataSource.postFriendData(
             request.toRequestDto(), page
